Move buttons, update comments
